commit via shell script

+++ Project2/part3.c
@@ -51,8 +51,12 @@ void alarm_handler(int signal){
     }
     pid_t w;
     int wstatus;
-    printf("ALARM: continuing child[%d]\n", pid_pool[running_child]);
-    kill(pid_pool[running_child], SIGCONT);
+    
+    if (w = waitpid(pid_pool[running_child], &wstatus, WNOHANG) == 0){
+      printf("ALARM: continuing child[%d]\n", pid_pool[running_child]);
+      kill(pid_pool[running_child], SIGCONT);
+    }
+
     sleep(4);
 
     if (w = waitpid(pid_pool[running_child], &wstatus, WNOHANG) == 0){
